Discount data type changed

+++ app/src/main/java/com/humaclab/selliscope/OrderActivity.java
@@ -342,7 +342,7 @@ public class OrderActivity extends AppCompatActivity implements View.OnClickList
                             EditText etDiscount = (EditText) row.findViewById(R.id.et_discount);
 
                             product.id = productID.get(sp.getSelectedItemPosition());
-                            product.discount = Integer.parseInt(etDiscount.getText().toString());
+                            product.discount = Double.parseDouble(etDiscount.getText().toString());
                             product.qty = Integer.parseInt(etQty.getText().toString());
                             product.row = Integer.parseInt(variantRows.get(i).get(0));
                             product.price = String.valueOf(productPrice.get(sp.getSelectedItemPosition()));
